# Speed up GroundPrimitive tests