@mramato 2018-11-08
This is excellent progress, but overall still way to slow.  Does the `largerScene` test need to exist at all?  These unit tests seem weird to me because they don't seem very unit-y.  I'm fine with taking an incremental change here but we should step back and re-evaluate `GroundPrimitiveSpec` as a whole instead of just trying to make what's there faster.  I'm guessing 17 seconds is like ~15% of all unit tests and that's just for 1 class?