@OmarShehata 2018-11-08
17 seconds is about 2.8% of all unit tests (total is 9 min and 50 sec on my machine).

Anything that is async definitely takes forever (in this case it's the topmost test, but that is specifically testing async). I think another source of slow down is web requests (`GoogleEarthEnterpriseMapsProvider` is now showing up as in the top 10 slowest, so I was going to look at that next). 

I don't have enough context to say whether we can ditch `largerScene` tests but also it's not the biggest slow down right now (and it's "large" in that it's a 2x2 canvas instead of 1x1).  

Like `renders in 3D` is 1 second and it doesn't do anything other than 2 `renderAndCall` functions. I think reading back the pixels is the bottleneck now. The only way I can see to get around that is to batch our tests so instead of having a lot of little draw calls and read pixels it'd be vastly faster to read once, but that's a much bigger architectural change.