@OmarShehata 2018-11-08
My hope is by systematically tackling all the slowest offenders (defined as > 1 second) I'll find some that can drop down to ms (like those external web requests are probably a good candidate) and that'll make a big enough dent.