@mramato 2018-11-08
> (total is 9 min and 50 sec on my machine).

!?!?! That's crazy slow.  Our target is under 60 seconds for the entire suite.  Running master on my machine now to see how far it's fallen.