@likangning93 2018-11-12
>  Works (or fails gracefully) in IE11.

I'm seeing a spec failure for `Scene/GroundPrimitive renders with invert classification and an opaque color`, but it's in `master` too and it looks like it was known in https://github.com/AnalyticalGraphicsInc/cesium/issues/2860#issuecomment-426016048. It's probably the same as the `Scene/ClassificationPrimitive renders with invert classification and an opaque color` failure in 1.51, but I'll make a note there anyway.

Shouldn't hold this PR back.