@OmarShehata 2018-11-15
@likangning93 cleaned up based on your feedback, and I think it actually saved a few seconds since we removed some calls from a `afterEach`. 

Also, my reported timings for the total were based on running it on the command line with `npm run test`. Running it in the browser with the spec runner is closer to the 3 minutes you guys report. I think the console print out might be skewing those timings, or something else there.